styl: sign페이지 파일 업데이트
